feat: improve LLM judge test infrastructure and fix test issues
- Move LLMJudge class to tests/helpers/llm-judge.ts as named export
- Fix test teardown to create test-results directory and use dynamic import for open()
- Improve fetch response handling with proper error checking and response normalization
- Update type definitions to include flags and conversationAnalysis fields
- Fix SSE tool_call branch to be schema-agnostic with better parameter handling
- Add biome-disable comments to test fixture exports
- Update README with clear instructions for running tests with wrangler dev
- Fix pattern matching in LLM judge tests for better response recognition
- All tests now passing (9/9) with improved reliability
